requirements.txt
Django==3.0.2
sorl-thumbnail==12.6.0
django-simple-captcha==0.5.12
django-precise-bbcode==1.2.10
django-cleanup==4.0.0